+++ Changelog.md
@@ -15,6 +15,11 @@ Bug Fixes:
   `have_received` matcher (they are not intended to be used together
   and previously caused an odd internal failure in rspec-mocks). (Jon Rowe, #788).
 
+Enhancements:
+
+* Treat `any_args` as an arg splat, allowing it to match an arbitrary
+  number of args at any point in an arg list. (Myron Marston, #786)
+
 ### 3.1.1 / 2014-09-18
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.1.0...v3.1.1)