+++ Changelog.md
@@ -1,14 +1,16 @@
-### Development
+### 3.1.3 Development
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.1.2...3-1-maintenance)
 
 Bug Fixes:
 
-* Correct received messages count when used with `have_received` matcher
+* Correct received messages count when used with `have_received` matcher.
   (Jon Rowe, #793)
 * Provide a clear error message when you use `allow_any_instance_of(...)` or
   `expect_any_instance_of(...)` with the `have_received` matcher (they are
   not intended to be used together and previously caused an odd internal
   failure in rspec-mocks). (Jon Rowe, #799).
+* Fix verified double `with` verification so that it applies to method
+  stubs. (Myron Marston, #790)
 
 ### 3.1.2 / 2014-09-26
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.1.1...v3.1.2)